Refactor http library
