+++ packages/compiler/core/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from "./projected-names-utils.js";
 export * from "./usage-resolver.js";
+export * from "./operation-utils.js";
