+++ packages/openapi3/src/openapi.ts
@@ -59,7 +59,7 @@ import {
 } from "@cadl-lang/openapi";
 import { Discriminator, getDiscriminator, http } from "@cadl-lang/rest";
 import {
-  getAllRoutes,
+  getAllHttpServices,
   getAuthentication,
   getContentTypes,
   getHeaderFieldName,
@@ -67,11 +67,11 @@ import {
   getQueryParamName,
   getStatusCodeDescription,
   HttpAuth,
+  HttpOperation,
   HttpOperationParameter,
   HttpOperationParameters,
   HttpOperationResponse,
   isStatusCode,
-  OperationDetails,
   reportIfNoRoutes,
   ServiceAuthentication,
 } from "@cadl-lang/rest/http";
@@ -297,10 +297,11 @@ function createOAPIEmitter(program: Program, options: ResolvedOpenAPI3EmitterOpt
   async function emitOpenAPIFromVersion(serviceNamespace: Namespace, version?: string) {
     initializeEmitter(serviceNamespace, version);
     try {
-      const [routes] = getAllRoutes(program);
-      reportIfNoRoutes(program, routes);
+      const [services] = getAllHttpServices(program);
+      const service = services[0];
+      reportIfNoRoutes(program, service.operations);
 
-      for (const operation of routes) {
+      for (const operation of service.operations) {
         emitOperation(operation);
       }
       emitReferences();
@@ -338,7 +339,7 @@ function createOAPIEmitter(program: Program, options: ResolvedOpenAPI3EmitterOpt
     }
   }
 
-  function emitOperation(operation: OperationDetails): void {
+  function emitOperation(operation: HttpOperation): void {
     const { path: fullPath, operation: op, verb, parameters } = operation;
 
     // If path contains a query string, issue msg and don't emit this endpoint
