+++ packages/rest/src/http/index.ts
@@ -1,6 +1,6 @@
 export const namespace = "Cadl.Http";
 
 export * from "./decorators.js";
+export * from "./operations.js";
 export * from "./responses.js";
-export * from "./route.js";
 export * from "./types.js";
