+++ packages/rest/src/http/route.ts
@@ -1,160 +1,25 @@
 import {
   createDiagnosticCollector,
-  DecoratorContext,
   Diagnostic,
-  DiagnosticCollector,
   filterModelProperties,
-  getServiceNamespace,
   Interface,
-  isGlobalNamespace,
-  isTemplateDeclaration,
-  isTemplateDeclarationOrInstance,
   ModelProperty,
   Namespace,
   Operation,
   Program,
   Type,
-  validateDecoratorTarget,
 } from "@cadl-lang/compiler";
 import { createDiagnostic, createStateSymbol, reportDiagnostic } from "../lib.js";
-import {
-  getAction,
-  getCollectionAction,
-  getResourceOperation,
-  getSegment,
-  getSegmentSeparator,
-  isAutoRoute,
-} from "../rest.js";
+import { getSegment, getSegmentSeparator, isAutoRoute } from "../rest.js";
 import { extractParamsFromPath } from "../utils.js";
 import {
   getHeaderFieldName,
-  getOperationVerb,
   getPathParamName,
   getQueryParamName,
-  HttpVerb,
+  getRoutePath,
   isBody,
 } from "./decorators.js";
-import { getResponsesForOperation, HttpOperationResponse } from "./responses.js";
-
-export type OperationContainer = Namespace | Interface;
-
-export interface FilteredRouteParam {
-  routeParamString?: string;
-  excludeFromOperationParams?: boolean;
-}
-
-export interface AutoRouteOptions {
-  routeParamFilter?: (op: Operation, param: ModelProperty) => FilteredRouteParam | undefined;
-}
-
-export interface RouteOptions {
-  autoRouteOptions?: AutoRouteOptions;
-}
-
-export interface HttpOperationParameter {
-  type: "query" | "path" | "header";
-  name: string;
-  param: ModelProperty;
-}
-
-export interface HttpOperationParameters {
-  parameters: HttpOperationParameter[];
-  bodyType?: Type;
-  bodyParameter?: ModelProperty;
-}
-
-export interface OperationDetails {
-  path: string;
-  pathFragment?: string;
-  verb: HttpVerb;
-  container: OperationContainer;
-  parameters: HttpOperationParameters;
-  responses: HttpOperationResponse[];
-  operation: Operation;
-}
-
-export interface RoutePath {
-  path: string;
-  isReset: boolean;
-}
-
-/**
- * `@route` defines the relative route URI for the target operation
- *
- * The first argument should be a URI fragment that may contain one or more path parameter fields.
- * If the namespace or interface that contains the operation is also marked with a `@route` decorator,
- * it will be used as a prefix to the route URI of the operation.
- *
- * `@route` can only be applied to operations, namespaces, and interfaces.
- */
-export function $route(context: DecoratorContext, entity: Type, path: string) {
-  setRoute(context, entity, {
-    path,
-    isReset: false,
-  });
-}
-
-export function $routeReset(context: DecoratorContext, entity: Type, path: string) {
-  setRoute(context, entity, {
-    path,
-    isReset: true,
-  });
-}
-
-const routeOptionsKey = createStateSymbol("routeOptions");
-export function setRouteOptionsForNamespace(
-  program: Program,
-  namespace: Namespace,
-  options: RouteOptions
-) {
-  program.stateMap(routeOptionsKey).set(namespace, options);
-}
-
-function getRouteOptionsForNamespace(
-  program: Program,
-  namespace: Namespace
-): RouteOptions | undefined {
-  return program.stateMap(routeOptionsKey).get(namespace);
-}
-
-const routesKey = createStateSymbol("routes");
-function setRoute(context: DecoratorContext, entity: Type, details: RoutePath) {
-  if (
-    !validateDecoratorTarget(context, entity, "@route", ["Namespace", "Interface", "Operation"])
-  ) {
-    return;
-  }
-
-  const state = context.program.stateMap(routesKey);
-
-  if (state.has(entity)) {
-    if (entity.kind === "Operation" || entity.kind === "Interface") {
-      reportDiagnostic(context.program, {
-        code: "duplicate-route-decorator",
-        messageId: entity.kind === "Operation" ? "operation" : "interface",
-        target: entity,
-      });
-    } else {
-      const existingValue: RoutePath = state.get(entity);
-      if (existingValue.path !== details.path) {
-        reportDiagnostic(context.program, {
-          code: "duplicate-route-decorator",
-          messageId: "namespace",
-          target: entity,
-        });
-      }
-    }
-  } else {
-    state.set(entity, details);
-  }
-}
-
-export function getRoutePath(
-  program: Program,
-  entity: Namespace | Interface | Operation
-): RoutePath | undefined {
-  return program.stateMap(routesKey).get(entity);
-}
+import { HttpOperationParameter, HttpOperationParameters, RouteOptions } from "./types.js";
 
 // The set of allowed segment separator characters
 const AllowedSegmentSeparators = ["/", ":"];
@@ -311,25 +176,20 @@ function generatePathFromParameters(
   addSegmentFragment(program, operation, pathFragments);
 }
 
-function getPathForOperation(
+export function resolvePathAndParameters(
   program: Program,
-  diagnostics: DiagnosticCollector,
   operation: Operation,
-  routeFragments: string[],
   options: RouteOptions
-): { path: string; pathFragment?: string; parameters: HttpOperationParameters } {
+): [{ path: string; parameters: HttpOperationParameters }, readonly Diagnostic[]] {
+  const diagnostics = createDiagnosticCollector();
   const parameters = diagnostics.pipe(getOperationParameters(program, operation));
-  const pathFragments = [...routeFragments];
-  const routePath = getRoutePath(program, operation);
+  let segments: string[] = [];
   if (isAutoRoute(program, operation)) {
     // The operation exists within an @autoRoute scope, generate the path.  This
     // mutates the pathFragments and parameters lists that are passed in!
-    generatePathFromParameters(program, operation, pathFragments, parameters, options);
+    generatePathFromParameters(program, operation, segments, parameters, options);
   } else {
-    // Prepend any explicit route path
-    if (routePath) {
-      pathFragments.push(routePath.path);
-    }
+    segments = getRouteSegments(program, operation);
 
     // Pull out path parameters to verify what's in the path string
     const paramByName = new Map(
@@ -339,7 +199,7 @@ function getPathForOperation(
     );
 
     // Find path parameter names used in all route fragments
-    const declaredPathParams = pathFragments.flatMap(extractParamsFromPath);
+    const declaredPathParams = segments.flatMap(extractParamsFromPath);
 
     // For each param in the declared path parameters (e.g. /foo/{id} has one, id),
     // delete it because it doesn't need to be added to the path.
@@ -361,120 +221,32 @@ function getPathForOperation(
 
     // Add any remaining declared path params
     for (const param of paramByName.keys()) {
-      pathFragments.push(`{${param}}`);
+      segments.push(`{${param}}`);
     }
   }
 
-  return {
-    path: buildPath(pathFragments),
-    pathFragment: routePath?.path,
+  return diagnostics.wrap({
+    path: buildPath(segments),
     parameters,
-  };
+  });
 }
 
-function getVerbForOperation(
-  program: Program,
-  diagnostics: DiagnosticCollector,
-  operation: Operation,
-  parameters: HttpOperationParameters
-): HttpVerb {
-  const resourceOperation = getResourceOperation(program, operation);
-  const verb =
-    (resourceOperation && resourceOperationToVerb[resourceOperation.operation]) ??
-    getOperationVerb(program, operation) ??
-    // TODO: Enable this verb choice to be customized!
-    (getAction(program, operation) || getCollectionAction(program, operation) ? "post" : undefined);
+function getRouteSegments(program: Program, target: Operation | Interface | Namespace): string[] {
+  const route = getRoutePath(program, target)?.path;
+  const seg = route ? [route] : [];
+  switch (target.kind) {
+    case "Namespace":
+      return target.namespace ? [...getRouteSegments(program, target.namespace), ...seg] : seg;
+    case "Interface":
+      return target.namespace ? [...getRouteSegments(program, target.namespace), ...seg] : seg;
 
-  if (verb !== undefined) {
-    return verb;
+    case "Operation":
+      return target.interface
+        ? [...getRouteSegments(program, target.interface), ...seg]
+        : target.namespace
+        ? [...getRouteSegments(program, target.namespace), ...seg]
+        : seg;
   }
-
-  // If no verb was found by this point, choose a verb based on whether there is
-  // a body type for the request
-  return parameters.bodyType ? "post" : "get";
-}
-
-function buildRoutes(
-  program: Program,
-  diagnostics: DiagnosticCollector,
-  container: OperationContainer,
-  routeFragments: string[],
-  visitedOperations: Set<Operation>,
-  options: RouteOptions
-): OperationDetails[] {
-  if (container.kind === "Interface" && isTemplateDeclaration(container)) {
-    // Skip template interface operations
-    return [];
-  }
-  // Get the route info for this container, if any
-  const baseRoute = getRoutePath(program, container);
-  const parentFragments = [...routeFragments, ...(baseRoute ? [baseRoute.path] : [])];
-
-  // TODO: Allow overriding the existing resource operation of the same kind
-
-  const operations: OperationDetails[] = [];
-  for (const [_, op] of container.operations) {
-    // Skip previously-visited operations
-    if (visitedOperations.has(op)) {
-      continue;
-    }
-
-    // Skip templated operations
-    if (isTemplateDeclarationOrInstance(op)) {
-      continue;
-    }
-
-    const route = getPathForOperation(program, diagnostics, op, parentFragments, options);
-    const verb = getVerbForOperation(program, diagnostics, op, route.parameters);
-    const responses = diagnostics.pipe(getResponsesForOperation(program, op));
-
-    operations.push({
-      path: route.path,
-      pathFragment: route.pathFragment,
-      verb,
-      container,
-      parameters: route.parameters,
-      operation: op,
-      responses,
-    });
-  }
-
-  // Build all child routes and append them to the list, but don't recurse in
-  // the global scope because that could pull in unwanted operations
-  if (container.kind === "Namespace") {
-    // If building routes for the global namespace we shouldn't navigate the sub namespaces.
-    const includeSubNamespaces = isGlobalNamespace(program, container);
-    const additionalInterfaces = getExternalInterfaces(program, container) ?? [];
-    const children: OperationContainer[] = [
-      ...(includeSubNamespaces ? [] : container.namespaces.values()),
-      ...container.interfaces.values(),
-      ...additionalInterfaces,
-    ];
-
-    const childRoutes = children.flatMap((child) =>
-      buildRoutes(program, diagnostics, child, parentFragments, visitedOperations, options)
-    );
-    for (const child of childRoutes) [operations.push(child)];
-  }
-
-  return operations;
-}
-
-export function getRoutesForContainer(
-  program: Program,
-  container: OperationContainer,
-  visitedOperations: Set<Operation>,
-  options?: RouteOptions
-): [OperationDetails[], readonly Diagnostic[]] {
-  const routeOptions =
-    options ??
-    (container.kind === "Namespace" ? getRouteOptionsForNamespace(program, container) : {}) ??
-    {};
-  const diagnostics = createDiagnosticCollector();
-
-  return diagnostics.wrap(
-    buildRoutes(program, diagnostics, container, [], visitedOperations, routeOptions)
-  );
 }
 
 const externalInterfaces = createStateSymbol("externalInterfaces");
@@ -497,104 +269,3 @@ export function includeInterfaceRoutesInNamespace(
 
   array.push(sourceInterface);
 }
-
-function getExternalInterfaces(program: Program, namespace: Namespace) {
-  const interfaces: string[] | undefined = program.stateMap(externalInterfaces).get(namespace);
-  if (interfaces === undefined) {
-    return undefined;
-  }
-  return interfaces
-    .map((interfaceFQN: string) => {
-      let current: Namespace | undefined = program.getGlobalNamespaceType();
-      const namespaces = interfaceFQN.split(".");
-      const interfaceName = namespaces.pop()!;
-      for (const namespaceName of namespaces) {
-        current = current?.namespaces.get(namespaceName);
-      }
-      return current?.interfaces.get(interfaceName);
-    })
-    .filter((x): x is Interface => x !== undefined);
-}
-
-export function getAllRoutes(
-  program: Program,
-  options?: RouteOptions
-): [OperationDetails[], readonly Diagnostic[]] {
-  let operations: OperationDetails[] = [];
-  const diagnostics = createDiagnosticCollector();
-  const serviceNamespace = getServiceNamespace(program);
-  const namespacesToExport: Namespace[] = serviceNamespace ? [serviceNamespace] : [];
-
-  const visitedOperations = new Set<Operation>();
-  for (const container of namespacesToExport) {
-    const newOps = diagnostics.pipe(
-      getRoutesForContainer(program, container, visitedOperations, options)
-    );
-
-    // Make sure we don't visit the same operations again
-    newOps.forEach((o) => visitedOperations.add(o.operation));
-
-    // Accumulate the new operations
-    operations = [...operations, ...newOps];
-  }
-
-  validateRouteUnique(diagnostics, operations);
-  return diagnostics.wrap(operations);
-}
-
-export function reportIfNoRoutes(program: Program, routes: OperationDetails[]) {
-  if (routes.length === 0) {
-    reportDiagnostic(program, {
-      code: "no-routes",
-      target: program.getGlobalNamespaceType(),
-    });
-  }
-}
-
-function validateRouteUnique(diagnostics: DiagnosticCollector, operations: OperationDetails[]) {
-  const grouped = new Map<string, Map<HttpVerb, OperationDetails[]>>();
-
-  for (const operation of operations) {
-    const { verb, path } = operation;
-    let map = grouped.get(path);
-    if (map === undefined) {
-      map = new Map<HttpVerb, OperationDetails[]>();
-      grouped.set(path, map);
-    }
-
-    let list = map.get(verb);
-    if (list === undefined) {
-      list = [];
-      map.set(verb, list);
-    }
-
-    list.push(operation);
-  }
-
-  for (const [path, map] of grouped) {
-    for (const [verb, routes] of map) {
-      if (routes.length >= 2) {
-        for (const route of routes) {
-          diagnostics.add(
-            createDiagnostic({
-              code: "duplicate-operation",
-              format: { path, verb, operationName: route.operation.name },
-              target: route.operation,
-            })
-          );
-        }
-      }
-    }
-  }
-}
-
-// TODO: Make this overridable by libraries
-const resourceOperationToVerb: any = {
-  read: "get",
-  create: "post",
-  createOrUpdate: "patch",
-  createOrReplace: "put",
-  update: "patch",
-  delete: "delete",
-  list: "get",
-};
