+++ packages/rest/src/http/types.ts
@@ -1,3 +1,14 @@
+import {
+  Interface,
+  ListOperationOptions,
+  ModelProperty,
+  Namespace,
+  Operation,
+  Type,
+} from "@cadl-lang/compiler";
+
+export type HttpVerb = "get" | "put" | "post" | "patch" | "delete" | "head";
+
 export interface ServiceAuthentication {
   /**
    * Either one of those options can be used independently to authenticate.
@@ -148,4 +159,97 @@ export interface ClientCredentialsFlow {
 export interface OAuth2Scope {
   value: string;
   description?: string;
+}
+
+export type OperationContainer = Namespace | Interface;
+
+export interface FilteredRouteParam {
+  routeParamString?: string;
+  excludeFromOperationParams?: boolean;
+}
+
+export interface AutoRouteOptions {
+  routeParamFilter?: (op: Operation, param: ModelProperty) => FilteredRouteParam | undefined;
+}
+
+export interface RouteOptions {
+  listOptions?: ListOperationOptions;
+  autoRouteOptions?: AutoRouteOptions;
+}
+
+export interface HttpOperationParameter {
+  type: "query" | "path" | "header";
+  name: string;
+  param: ModelProperty;
+}
+
+export interface HttpOperationParameters {
+  parameters: HttpOperationParameter[];
+  bodyType?: Type;
+  bodyParameter?: ModelProperty;
+}
+
+export interface HttpService {
+  namespace: Namespace;
+  operations: HttpOperation[];
+}
+
+/**
+ * @deprecated
+ */
+export type OperationDetails = HttpOperation;
+
+export interface HttpOperation {
+  /**
+   * Route path
+   */
+  path: string;
+
+  /**
+   * Route verb.
+   */
+  verb: HttpVerb;
+
+  /**
+   * Parent type being the interface, namespace or global namespace.
+   */
+  container: OperationContainer;
+
+  /**
+   * Parameters.
+   */
+  parameters: HttpOperationParameters;
+
+  /**
+   * Responses.
+   */
+  responses: HttpOperationResponse[];
+
+  /**
+   * Operation type reference.
+   */
+  operation: Operation;
+}
+
+export interface RoutePath {
+  path: string;
+  isReset: boolean;
+}
+
+export type StatusCode = `${number}` | "*";
+export interface HttpOperationResponse {
+  statusCode: StatusCode;
+  type: Type;
+  description?: string;
+  responses: HttpOperationResponseContent[];
+}
+
+export interface HttpOperationResponseContent {
+  headers?: Record<string, ModelProperty>;
+  body?: HttpOperationBody;
+}
+
+export interface HttpOperationBody {
+  contentTypes: string[];
+  type: Type;
 }
